pyproject.toml
[tool.poetry]
name = "automationkit"
description = "Automation Kit Test Framework"
version = "0.3.4"
authors = [
    "Myron Walker <myron.walker@gmail.com>"
]
readme = "README.md"
license = "LICENSE.txt"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX"
]
keywords = [
    "automation",
    "continuous-integration",
    "distributed-test-framework",
    "distributed-testing-tools",
    "distributed-testing",
    "upnp",
    "test-automation",
    "test-framework",
    "testing",
    "testing-tools"
]
packages = [{include = "akit", from="source/packages"}]

[tool.poetry.dependencies]
python = ">=3.8,<4.0"
rope = "^1.6.0"
debugpy = "^1.6.5"
click = "^8.1.3"
coverage = "^7.0.4"
ipython = "^8.8.0"
netifaces = "^0.11.0"
pyserial = "^3.5"
pycrunch-trace = "^0.2"
paramiko = "^2.12.0"
pylint = "^2.15.9"
pyreadline3 = "^3.4.1"
requests = "^2.28.1"
zeroconf = "^0.47.1"
rich = "^13.0.1"
ssdp = "^1.1.1"
pyyaml = "^6.0"
dlipower = "^1.0.176"
gunicorn = "^20.1.0"
werkzeug = "^2.2.2"
flask = "^2.2.2"
flask-migrate = "^4.0.1"
flask-restx = "^1.0.3"

[tool.poetry.group.dev.dependencies]
cython = "^0.29.33"
myst-parser = "^0.18.1"
sphinx = ">=1.6,<6"
sphinx-rtd-theme = "^1.1.1"
sqlalchemy = "^1.4.46"
sqlalchemy-utils = "^0.39.0"


[tool.poetry.group.dbio.dependencies]
psycopg2 = "^2.9.5"
sqlalchemy = "^1.4.46"
sqlalchemy-utils = "^0.39.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
